fix: set permissions on the model-server when clicking "Model History"
The workspace client uses an access token generated by the workspace-manager to get access to the
repository, but when clicking on "Model History" the user directly works on the model-server with
his identity token generated by the OAuth proxy. The workspace-manager now modifies the permissions
on the model-server before forwarding the user to it.

Ideally, we would have a common authorization service than manages the permission for all services,
but that more effort.