feat: workspace containers are only started after a successful build
Previously, it could happen that, if the cluster was low on resources, the workspace instance was
blocking the resources required for starting the build job.

Also, if a build job fails the progress bar jumps back to 0% and a message
"Workspace build failed" is shown with a link to the build log.

Previously, the progress bar was just stuck and the user waited without knowing that there is
something wrong with the configuration.